+++ services/explorer-ui/src/pages/tx-effect-details/index.tsx
@@ -2,7 +2,7 @@ import { useParams } from "@tanstack/react-router";
 import { type FC } from "react";
 import { DroppedBanner } from "~/components/dropped-banner";
 import { KeyValueDisplay } from "~/components/info-display/key-value-display";
-import { LoadingDetails } from "~/components/loading/tx-effect";
+import { LoadingDetails } from "~/components/loading/loading-details";
 import { getEmptyTxEffectData } from "~/components/loading/util";
 import { OrphanedBanner } from "~/components/orphaned-banner";
 import {
@@ -86,7 +86,12 @@ export const TxEffectDetails: FC = () => {
   // No data found
   // TODO: Make proper error page when no transaction is found
   if (!txEffects) {
-    return <div>No transaction found. (Perhaps it has just not reached our indexer yet, that might take a couple of seconds.)</div>;
+    return (
+      <div>
+        No transaction found. (Perhaps it has just not reached our indexer yet,
+        that might take a couple of seconds.)
+      </div>
+    );
   }
 
   // Success state - data is available
